+++ SVDModel/SVDCore/core/climate.h
@@ -15,7 +15,7 @@ public:
     // access
     /// retrieve a list of climate series, starting from 'start_year' (first year: 1, ...)
     /// and with the given length ('series_length').
-    std::vector< const std::vector<float>* > series(int start_year, int series_length, int climateId) const;
+    std::vector< const std::vector<float>* > series(int start_year, size_t series_length, int climateId) const;
     const std::vector<float> &singleSeries(const int year, const int climateId) const { return mData.at(year).at(climateId); }
     bool hasSeries(const int year, const int climateId) const { auto y=mData.find(year); if(y==mData.end()) return false;
                                                                 auto s= y->second.find(climateId); if (s==(*y).second.end()) return false;
